Refactored regular_subset().
