+++ phy/utils/array.py
@@ -251,6 +251,22 @@ def get_excerpts(data, n_excerpts=None, excerpt_size=None):
                                                  excerpt_size=excerpt_size)])
 
 
+def regular_subset(spikes=None, n_spikes_max=None):
+    """Prune the current selection to get at most n_spikes_max spikes."""
+    assert spikes is not None
+    # Nothing to do if the selection already satisfies n_spikes_max.
+    if n_spikes_max is None or len(spikes) <= n_spikes_max:
+        return spikes
+    step = int(np.clip(1. / n_spikes_max * len(spikes),
+                       1, len(spikes)))
+    # Random shift.
+    start = np.random.randint(low=0, high=step)
+    my_spikes = spikes[start::step][:n_spikes_max]
+    assert len(my_spikes) <= len(spikes)
+    assert len(my_spikes) <= n_spikes_max
+    return my_spikes
+
+
 # -----------------------------------------------------------------------------
 # PartialArray
 # -----------------------------------------------------------------------------
